pyproject.toml
[tool.poetry]
name = "desid-playground"
version = "0.1.0"
description = "DESID Data Analytics Platform"
authors = ["Fred Guth <frederico.guth@saude.gov.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10"
sqlmesh = {extras = ["web"], version = "^0.128.0"}
duckdb = "^1.0.0"
numpy = "^2.0.1"
pandas = "^2.2.2"
ipykernel = "^6.29.5"
cx-oracle = "^8.3.0"
oracledb = "^2.4.1"
sqlalchemy = "^2.0.35"
fastduck = "^0.0.14"
tqdm = "^4.66.5"
ipywidgets = "^8.1.5"
fastparquet = "^2024.5.0"
pyarrow = "^18.0.0"
filelock = "^3.16.1"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
notebook = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
